move release undraft step until after canary is fully published
