+++ scripts/publish-release.js
@@ -86,6 +86,75 @@ const cwd = process.cwd()
     await publish(pkg, retry + 1)
   }
 
+  const undraft = async () => {
+    const githubToken = process.env.RELEASE_BOT_GITHUB_TOKEN
+
+    if (!githubToken) {
+      throw new Error(`Missing RELEASE_BOT_GITHUB_TOKEN`)
+    }
+
+    if (isCanary) {
+      try {
+        const ghHeaders = {
+          Accept: 'application/vnd.github+json',
+          Authorization: `Bearer ${githubToken}`,
+          'X-GitHub-Api-Version': '2022-11-28',
+        }
+        const { version: _version } = require('../lerna.json')
+        const version = `v${_version}`
+
+        let release
+        let releasesData
+
+        // The release might take a minute to show up in
+        // the list so retry a bit
+        for (let i = 0; i < 6; i++) {
+          try {
+            const releaseUrlRes = await fetch(
+              `https://api.github.com/repos/vercel/next.js/releases`,
+              {
+                headers: ghHeaders,
+              }
+            )
+            releasesData = await releaseUrlRes.json()
+
+            release = releasesData.find(
+              (release) => release.tag_name === version
+            )
+          } catch (err) {
+            console.log(`Fetching release failed`, err)
+          }
+          if (!release) {
+            console.log(`Retrying in 10s...`)
+            await new Promise((resolve) => setTimeout(resolve, 10 * 1000))
+          }
+        }
+
+        if (!release) {
+          console.log(`Failed to find release`, releasesData)
+          return
+        }
+
+        const undraftRes = await fetch(release.url, {
+          headers: ghHeaders,
+          method: 'PATCH',
+          body: JSON.stringify({
+            draft: false,
+            name: version,
+          }),
+        })
+
+        if (undraftRes.ok) {
+          console.log('un-drafted canary release successfully')
+        } else {
+          console.log(`Failed to undraft`, await undraftRes.text())
+        }
+      } catch (err) {
+        console.error(`Failed to undraft release`, err)
+      }
+    }
+  }
+
   await Promise.allSettled(
     packageDirs.map(async (packageDir) => {
       const pkgJson = JSON.parse(
@@ -102,4 +171,6 @@ const cwd = process.cwd()
       await publish(packageDir)
     })
   )
+
+  await undraft()
 })()
