+++ scripts/start-release.js
@@ -2,7 +2,6 @@ const path = require('path')
 const execa = require('execa')
 const resolveFrom = require('resolve-from')
 const ansiEscapes = require('ansi-escapes')
-const fetch = require('node-fetch')
 
 function getPromptErrorDetails(rawAssertion, mostRecentChunk) {
   const assertion = rawAssertion.toString().trim()
@@ -162,65 +161,6 @@ async function main() {
   console.log('Await child process...')
   await child
   console.log('Release process is finished')
-
-  if (isCanary) {
-    try {
-      const ghHeaders = {
-        Accept: 'application/vnd.github+json',
-        Authorization: `Bearer ${githubToken}`,
-        'X-GitHub-Api-Version': '2022-11-28',
-      }
-      let { version } = require('../lerna.json')
-      version = `v${version}`
-
-      let release
-      let releasesData
-
-      // The release might take a minute to show up in
-      // the list so retry a bit
-      for (let i = 0; i < 6; i++) {
-        try {
-          const releaseUrlRes = await fetch(
-            `https://api.github.com/repos/vercel/next.js/releases`,
-            {
-              headers: ghHeaders,
-            }
-          )
-          releasesData = await releaseUrlRes.json()
-
-          release = releasesData.find((release) => release.tag_name === version)
-        } catch (err) {
-          console.log(`Fetching release failed`, err)
-        }
-        if (!release) {
-          console.log(`Retrying in 10s...`)
-          await new Promise((resolve) => setTimeout(resolve, 10 * 1000))
-        }
-      }
-
-      if (!release) {
-        console.log(`Failed to find release`, releasesData)
-        return
-      }
-
-      const undraftRes = await fetch(release.url, {
-        headers: ghHeaders,
-        method: 'PATCH',
-        body: JSON.stringify({
-          draft: false,
-          name: version,
-        }),
-      })
-
-      if (undraftRes.ok) {
-        console.log('un-drafted canary release successfully')
-      } else {
-        console.log(`Failed to undraft`, await undraftRes.text())
-      }
-    } catch (err) {
-      console.error(`Failed to undraft release`, err)
-    }
-  }
 }
 
 main()
